Fix bug caused by files without extensions
the initial call to get_ext uses split_string to find everything after the last "." . If no "." was present in the string, the returned result would be the entire filename. This would cause a false positive for a file with the same name as one of the blacklisted extensions. For example, a file with the name "bat".